changed calling convention of fprops_solve_ph, fprops_solve_px, fprops_solve_Tx. added testing of viscosity for sodium (incompressible) part-way through implementing support for incompressible fluids in python/ascend interfaces.
git-svn-id: https://svn.ascend4.org/branches/fprops-incomp@3431 f15cd148-d149-4907-926e-73cf92b3ee03

+++ models/johnpye/fprops/filedata.h
@@ -435,7 +435,7 @@ typedef struct ViscDataEpt_struct{
     double mu_star;
     unsigned np;
     ViscPowTerm *pt; ///< viscosity terms c*T^t
-    unsigned b; ///< viscosity term b*ln(T)
+    double b; ///< viscosity term b*ln(T)
     char is_ln; ///< if true, ln(mu) = [...]; if false, mu = [...].
 } ViscDataEpt;
